Add fix for missing process list in AS002
Occasionally, for reasons unknown and in a very small number of
cases, the 'ss' command usage in agent_status_check.AS002 fails
to return the 'users:((' section which lists which process is
using the 8142 socket.

This patch adjusts the check so that if the 'users:(())' section
data is returned as expected then the check continues to rely
on it. However, if it is not returned then the check will fall
back to reporting success based on if there is a connection to
8142 or not, similarly to the Windows check.