Extend the AS002 timeout value
On systems with many thousands of processes the use of the "-p"
option to resolve the process owning the connection takes a few
seconds as each process in the process table has to be queried
for the associated sockets to make the socket/process match.

Based on some limited testing, doubling the timeout from 2 to
4 seconds for this command should allow for success on moderately
busy systems with somewhere in the tens of thousands of processes.